2211: Add statement link and backup managers to 12-week test UI
