+++ accessibility_monitoring_platform/apps/audits/templates/audits/forms/twelve_week_statement_link.html
@@ -1,0 +1,21 @@
+{% extends 'common/case_form.html' %}
+
+{% block preform %}
+<div class="govuk-grid-row amp-margin-bottom-30">
+    <div class="govuk-grid-column-full">
+        {% include "audits/helpers/statement_links_backups.html" %}
+    </div>
+</div>
+{% endblock %}
+
+{% block form %}
+    {% include "common/messages.html" %}
+    <form method="post" action="{% url sitemap.current_platform_page.url_name audit.id %}">
+        {% csrf_token %}
+        {% include 'common/form_hidden_fields.html' with hidden_fields=form.hidden_fields %}
+        {% include 'common/form_hidden_fields.html' with hidden_fields=statement_link_form.hidden_fields %}
+        {% include 'common/amp_field.html' with field=statement_link_form.statement_url %}
+        {% include 'common/amp_field.html' with field=form.audit_retest_statement_pages_complete_date %}
+        {% include 'simplified/helpers/save_continue_cancel.html' %}
+    </form>
+{% endblock %}
