+++ accessibility_monitoring_platform/apps/audits/templates/audits/helpers/statement_links_backups.html
@@ -9,6 +9,8 @@
         </li>
         {% if statement_page.created %}
             <li>Added {{ statement_page.created|amp_datetime }}</li>
+        {% else %}
+            <li>Added 1 January 1970</li>
         {% endif %}
     </ul>
 {% endfor %}
