+++ stack_tests/integration_tests/cypress/e2e/new_simplified_case_test_report.cy.js
@@ -61,7 +61,7 @@ describe('Create simplified case, tests and report', () => {
     cy.title().should('eq', `${newDomain} · ${newOrganisationName} | Statement links manager`)
     cy.contains('Save and continue').click()
 
-    cy.title().should('eq', `${newDomain} · ${newOrganisationName} | Statement backups`)
+    cy.title().should('eq', `${newDomain} · ${newOrganisationName} | Statement backups manager`)
     cy.contains('Save and continue').click()
 
     cy.title().should('eq', `${newDomain} · ${newOrganisationName} | Statement overview`)
@@ -177,7 +177,10 @@ describe('Create simplified case, tests and report', () => {
     cy.title().should('eq', `${newDomain} · ${newOrganisationName} | WCAG summary`)
     cy.contains('Save and continue').click()
 
-    cy.title().should('eq', `${newDomain} · ${newOrganisationName} | Statement links`)
+    cy.title().should('eq', `${newDomain} · ${newOrganisationName} | Statement links manager`)
+    cy.contains('Save and continue').click()
+
+    cy.title().should('eq', `${newDomain} · ${newOrganisationName} | Statement backups manager`)
     cy.contains('Save and continue').click()
 
     cy.title().should('eq', `${newDomain} · ${newOrganisationName} | Statement overview`)
